+++ ricaun.Revit.ExtensibleJson.Tests/RevitTest.cs
@@ -1,0 +1,25 @@
+﻿using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.DB;
+using NUnit.Framework;
+using System;
+
+namespace ricaun.Revit.ExtensibleJson.Tests
+{
+    public class RevitTests
+    {
+        protected Application application;
+
+        [OneTimeSetUp]
+        public void Setup(Application application)
+        {
+            this.application = application;
+        }
+
+        [Test]
+        public void RevitTests_VersionName()
+        {
+            Assert.IsNotNull(application);
+            Console.WriteLine(application.VersionName);
+        }
+    }
+}